További logolás, és névrészlet felismerés teszt

+++ FunctionsCore/Models/DeliveryModel.cs
@@ -21,7 +21,6 @@ namespace FunctionsCore.Models
         public byte[] ValueBytes { get; set; }
         public Nyelvek ValueNyelv { get; set; }
         public bool SendedFl { get; set; } = false;
-
-        public int NumberOfSending { get; set; } = 0;
+        public byte NumberOfSending { get; set; } = 0;
     }
 }
